split draw box and pointer into their own files. so when i get to redoing lgsg and lobbys those functions won't be lost. because all of the menus use them.
